bugfix: selectedIndex to int -> long

+++ Classes/parts/DialogSelectListViewLayer.h
@@ -15,7 +15,7 @@ class DialogSelectListViewLayer : public cocos2d::Layer
 {
     
 public:
-    typedef std::function<void(int selectedIndex)> DialogSelectListener;
+    typedef std::function<void(long selectedIndex)> DialogSelectListener;
     
     DialogSelectListViewLayer();
     virtual ~DialogSelectListViewLayer();
@@ -33,7 +33,7 @@ public:
 protected:
     
 private:
-    int _selectedIndex;
+    long _selectedIndex;
     cocos2d::Node* _baseLayout;
     
     DialogSelectListener _okListener;
